MTKHelpers

Documentation for MTKHelpers.

MTKHelpers.ProblemParSetter

MTKHelpers.ProblemParSetter

MTKHelpers.count_par

MTKHelpers.count_paropt

MTKHelpers.count_states

MTKHelpers.count_states

MTKHelpers.embed_system

MTKHelpers.get_paropt

MTKHelpers.get_paropt

MTKHelpers.label_par

MTKHelpers.label_paropt

MTKHelpers.label_state

MTKHelpers.label_state

MTKHelpers.paroptsyms

MTKHelpers.parsyms

MTKHelpers.statesyms

MTKHelpers.statesyms

MTKHelpers.strip_namespace

MTKHelpers.symbol

MTKHelpers.update_statepar

MTKHelpers.update_statepar

MTKHelpers.ProblemParSetter — Method

ProblemParSetter(state_names,par_names,popt_names) 
ProblemParSetter(sys::ODESystem, popt_names; strip=false)

Helps keeping track of a subset of initial states and paraemters to be optimized.

Arguments

state_names: Tuple or Vector of all the initial states of the problem

par_names: all the parameters of the problem

popt_names: the parameter/initial states to be optimized.

The states and parameters can be extracted from an ModelingToolkit.ODESystem. If strip=true, then namespaces of parameteres of a composed system are removed, e.g. subcomp₊p becomes p.

MTKHelpers.count_par — Method

count_states(::ProblemParSetter) 
count_par(::ProblemParSetter) 
count_paropt(::ProblemParSetter)

Report the number of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.count_paropt — Method

count_states(::ProblemParSetter) 
count_par(::ProblemParSetter) 
count_paropt(::ProblemParSetter)

Report the number of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.count_states — Method

count_states(::ProblemParSetter) 
count_par(::ProblemParSetter) 
count_paropt(::ProblemParSetter)

Report the number of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.embed_system — Method

embed_system(m;name)

Embeds system m as the single component of a larger system.

Then the naming of the states, parameters, observables matches the namespace of the system.

#using DifferentialEquations, Plots
function samplesystem(;name) 
@variables t 
D = Differential(t) 
sts = @variables x(t) RHS(t)  # RHS is observed
ps = @parameters τ       # parameters
ODESystem([ RHS  ~ (1 - x)/τ, D(x) ~ RHS ], t, sts, ps; name)
end                     
@named m = samplesystem()
@named sys = embed_system(m)
# note that state key is m.x, hence only m needs to be defined
prob = ODEProblem(sys, [m.x => 0.0], (0.0,10.0), [m.τ => 3.0])
sol = solve(prob);
plot(sol, vars=[m.x,m.RHS])    
plot(sol, vars=getproperty.(Ref(m),[:x, :RHS])) # access by symbols

MTKHelpers.get_paropt — Method

get_paropt(ps::ProblemParSetter, prob::ODEProblem; kwargs...)
get_paropt(ps::ProblemParSetter, u0, p; label=Val(false))

Extract the initial states and parameters corresponding to the positions that are optimized. If both u0 and p are vectors, the result is a vector, otherwise the result is a Tuple.

If label=Val(true), then label_paropt (see label_state) is called on the return value. The default is label=Val(false) to return plain Vectors to work smoothly with Optimizers.

MTKHelpers.label_par — Method

label_state(ps, u::Vector) = LArray{statesyms(ps)}(u)
label_par(ps, par::Vector) = LArray{parsyms(ps)}(par)
label_paropt(ps, popt::Vector) = LArray{paroptsyms(ps)}(popt)

Produce a labeled version of a sequence of initial states, parameters, or optimized parameters respectively. The return type differs given the input

SVector -> SLVector

NTuple -> NamedTuple

Vector -> LArray

MTKHelpers.label_paropt — Method

label_state(ps, u::Vector) = LArray{statesyms(ps)}(u)
label_par(ps, par::Vector) = LArray{parsyms(ps)}(par)
label_paropt(ps, popt::Vector) = LArray{paroptsyms(ps)}(popt)

Produce a labeled version of a sequence of initial states, parameters, or optimized parameters respectively. The return type differs given the input

SVector -> SLVector

NTuple -> NamedTuple

Vector -> LArray

MTKHelpers.label_state — Method

label_state(ps, u::Vector) = LArray{statesyms(ps)}(u)
label_par(ps, par::Vector) = LArray{parsyms(ps)}(par)
label_paropt(ps, popt::Vector) = LArray{paroptsyms(ps)}(popt)

Produce a labeled version of a sequence of initial states, parameters, or optimized parameters respectively. The return type differs given the input

SVector -> SLVector

NTuple -> NamedTuple

Vector -> LArray

MTKHelpers.paroptsyms — Method

statesyms(ps::ProblemParSetter)
parsyms(ps::ProblemParSetter)
paroptsyms(ps::ProblemParSetter)

Report the names, i.e. symbols of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.parsyms — Method

statesyms(ps::ProblemParSetter)
parsyms(ps::ProblemParSetter)
paroptsyms(ps::ProblemParSetter)

Report the names, i.e. symbols of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.statesyms — Method

statesyms(ps::ProblemParSetter)
parsyms(ps::ProblemParSetter)
paroptsyms(ps::ProblemParSetter)

Report the names, i.e. symbols of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.strip_namespace — Method

Omit the part before the first dot

MTKHelpers.symbol — Method

Extract the basic symbol

MTKHelpers.update_statepar — Method

prob_new = update_statepar(ps::ProblemParSetter, popt, prob::ODEProblem) 
u0new, pnew = update_statepar(ps::ProblemParSetter, popt, u0, p)

Return an updated problem or updates states and parameters where values corresponding to positions in popt are set.