CairoMakie plotting utils

MTKHelpers.pdf_figure — Function

pdf_figure = (size_inches = cm2inch.((8.3,8.3/1.618)); fontsize=9, kwargs...)

Creates a figure and Axis with specified resolution and fontsize for given figure size. See also cm2inch and save.

MTKHelpers.series_sol! — Function

series_sol!(ax, sol::AbstractODESolution, vars; tspan=extrema(sol.t), labels=string.(vars), nt=120, kwargs...)

calls CairoMakie.series for a grid fo n points and interpolated values from sol. Currently works only with solutions created by a non-composite solver, e.g. Tsit5.