Representing a system as a component

MTKHelpers.embed_system — Function

embed_system(m;name)

Embeds system m as the single component of a larger system. This helps to match the naming of the states, parameters, observables to the namespace of the system.

using ModelingToolkit, DifferentialEquations
using MTKHelpers
function samplesystem(;name) 
    @variables t 
    D = Differential(t) 
    sts = @variables x(t) RHS(t)  # RHS is observed
    ps = @parameters τ       # parameters
    ODESystem([ RHS  ~ (1 - x)/τ, D(x) ~ RHS ], t, sts, ps; name)
end                     
@named m = samplesystem()
@named sys = embed_system(m)

# note that keys are `m.x`,`m.τ` or `m.RHS`.
# Hence only m needs to be defined rather than all the states, parameters, 
# and observables of the system.
prob = ODEProblem(sys, [m.x => 0.0], (0.0,10.0), [m.τ => 3.0])
sol = solve(prob);
dx1 = sol[m.RHS][1:3] 
dx2 = sol[getproperty(m,:RHS)][1:3]  # access by symbols

# using Plots
# plot(sol, vars=[m.x,m.RHS])    
# plot(sol, vars=getproperty.(Ref(m),[:x, :RHS])) # access by symbols
length(dx2) == 3
# output
true

Utilities

MTKHelpers.symbol — Function

symbol(t)

Extract the inner symbol from a Term or Num object.

MTKHelpers.strip_namespace — Function

strip_namespace(s)

Omit the part before the first dot.

MTKHelpers.statesyms — Method

statesyms(sys::ODYSystem)
parsyms(sys::ODYSystem)

Extract the basic symbols without namespace of system states and system parameters.

MTKHelpers.cm2inch — Function

cm2inch(x)

Convert length in cm to inch units: 1 inch = 2.54 cm.