Setting Initial state and Parameters of a problem

Often one wants to change a subset of the initial states,u0, and a subset of parameters,p, of an ODEProblem during an optimization.

Given u0 and p can be expressed as ComponentVectors, and popt can be expressed as a ComponentVector of optimized parameters, which may include initial states, the following class helps updating the corresponding positions in an ODEProblem.

Initial states and parameter components must have different names.

# setting up a simple example composite system and problem
using ModelingToolkit, DifferentialEquations, ComponentArrays
using MTKHelpers
function samplesystem(;name,τ = 3.0, p1=1.1, p2=1.2)
    @variables t
    D = Differential(t)
    sts = @variables x(t) RHS(t)             # RHS is observed
    ps = @parameters τ=τ p1=p1 p2=p2       # parameters
    ODESystem([ RHS  ~ p1/p2 * (1 - x)/τ, D(x) ~ RHS ], t, sts, ps; name)
end
@named m = samplesystem()
@named sys = embed_system(m)
prob = ODEProblem(sys, [m.x => 0.0], (0.0,10.0))

# setup position matching
popt = ComponentVector(m₊x=0.1, m₊p1=2.1)
pset = ProblemParSetter(sys, popt) # not fully type-inferred

# extract optimized
get_paropt(pset, prob)          # plain vector
get_paropt_labeled(pset, prob)  # ComponentVector
name_paropt(pset, prob)         # NamedVector

# update states and parameters
prob2 = update_statepar(pset, popt, prob)
prob2.p # p is still a plain vector
label_par(pset, prob2.p).m₊p1 == popt.m₊p1 # attach labels and access properties
label_state(pset, prob2.u0).m₊x == popt.m₊x # attach labels and access properties
get_paropt_labeled(pset, prob2) == popt

true

Note that ProblemParSetter, pset, is only fully type-inferred when constructed with three ComponentArrays.Axis objects. This propagates to all ComponentVectors constructed by it, e.g. with label_state. Hence, its recommended to pass pset across a function barrier for code where performance matters.

ProblemUpdater

A ProblemParSetter can be combined with a KeysProblemParGetter or other specific implementations of AbstractProblemParGetter to update an ODEProblem based on information already present in the ODEProblem.

The following example updates parameters k_R and k_P in the ODEProblem to the value of k_L. This can be useful to ensure that these parameters are also changed when optimizing parameter k_L.

An implementations of AbstractProblemParGetter can use any computation of the source keys to provide its destination keys. It should implement the keys method, so that when constructing the ProblemUpdater, consistent keys are used, as in the example below.

f = (u,p,t) -> p[1]*u
u0 = (L=1/2,)
p = (k_L = 1.0, k_R = 2.0, k_P = 3.0)
tspan = (0.0,1.0)
prob = ODEProblem(f,collect(u0),tspan,collect(p))

source = (:k_L,:k_L)
dest = (:k_R,:k_P)
pu = ProblemUpdater(KeysProblemParGetter(source, dest), keys(u0), keys(p))
prob2 = pu(prob)
p2 = label_par(par_setter(pu), prob2.p)
p2.k_R == p.k_L
p2.k_P == p.k_L

true

MTKHelpers.ProblemUpdater — Type

ProblemUpdater(par_getter, par_setter) 
ProblemUpdater(par_getter, u0_keys, p_keys)

Encapsulates updating an ODEProblem based on the problem itself by Callable (pu::ProblemUpdater)(prob).

Must be initialized with a callable AbstractProblemParGetter, e.g. KeysProblemParGetter and on a AbstractProblemParSetter, e.g. ProblemParSetter.

The second form of the constructor creates a ProblemParSetter based on keys(par_getter).

ProblemParGetter

MTKHelpers.KeysProblemParGetter — Type

KeysProblemParGetter(source_keys::NTuple{N,Symbol})

Provices callable (pg::KeysProblemParGetter)(pu::ProblemUpdater, prob)]. To be used to get the parameters/state vector to be set by ProblemUpdater.

Initialize with an NTuple of symbols that index into vcat(label_state(pu.pset, prob.u0), label_par(pu.pset, prob.p)).

In order to provide computations for parameters to set, declare a concrete subtype of KeysProblemParGetter, and implement a custom method (pg::MyProblemParGetter)(pu::ProblemUpdater, prob) that returns a vector of parameter values.

ProblemParSetter

The following type stores the necessary information, that can be queried.

MTKHelpers.ProblemParSetter — Type

ProblemParSetter(state_names,par_names,popt_names) 
ProblemParSetter(sys::ODESystem, popt_names; strip=false)

Helps keeping track of a subset of initial states and paraemters to be optimized.

Arguments

state_names: ComponentVector or Axis of all the initial states of the problem

par_names: all the parameters of the problem

popt_names: the parameter/initial states to be optimized.

If all of state_names, par_names, and popt_names are type-inferred Axes, then also the constructed ProblemParSetter is type-inferred.

The states and parameters can be extracted from an ModelingToolkit.ODESystem. If strip=true, then namespaces of parameters of a composed system are removed, e.g. subcomp₊p becomes p.

MTKHelpers.count_state — Function

count_state(::AbstractProblemParSetter) 
count_par(::AbstractProblemParSetter) 
count_paropt(::AbstractProblemParSetter)

Report the number of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.axis_state — Function

axis_state(pset::AbstractProblemParSetter)
axis_par(pset::AbstractProblemParSetter)
axis_paropt(pset::AbstractProblemParSetter)

Report the Axis, i.e. nested component symbols of problem states, problem parameters and optimized parameters respectively. Returns an AbstractAxis.

MTKHelpers.symbols_state — Function

symbols_state(pset::AbstractProblemParSetter)
symbols_par(pset::AbstractProblemParSetter)
symbols_paropt(pset::AbstractProblemParSetter)

Report the names, i.e. symbols of problem states, problem parameters and optimized parameters respectively, i.e. the concatenation of components. Similar to ComponentArrays.label, but inferred from Axis object.

Updating a problem

MTKHelpers.update_statepar — Method

prob_new = update_statepar(ps::ProblemParSetter, popt, prob::ODEProblem) 
u0new, pnew = update_statepar(ps::ProblemParSetter, popt, u0, p)

Return an updated problem or updates states and parameters where values corresponding to positions in popt hold the values of popt. The type is changed to the promotion type of popt, to allow working with dual numbers.

Extracting optimized parameters

MTKHelpers.get_paropt — Method

get_paropt(pset::AbstractProblemParSetter, prob::ODEProblem; kwargs...)
get_paropt(pset::AbstractProblemParSetter, u0, p)

get_paropt_labeled(pset::AbstractProblemParSetter, prob::ODEProblem; kwargs...)
get_paropt_labeled(pset::AbstractProblemParSetter, u0, p)

Extract the initial states and parameters corresponding to the positions that are optimized. If both u0 and p are AbstractVectors, the result is a Vector, otherwise the result is a Tuple.

The labeled versions additionally call label_paropt (see label_state) on the return value.

Labeling

MTKHelpers.label_state — Method

label_state(pset::AbstractProblemParSetter, u::AbstractVector) 
label_par(pset::AbstractProblemParSetter, par::AbstractVector) 
label_paropt(pset::AbstractProblemParSetter, popt::AbstractVector)

Produce a labeled version, i.e. a ComponentVector of initial states, parameters, or optimized parameters respectively.

MTKHelpers.name_state — Method

name_state(pset, u::AbstractVector) 
name_par(pset, par::AbstractVector) 
name_paropt(pset, popt::AbstractVector) 

name_paropt(pset::AbstractProblemParSetter, prob::ODEProblem)

Produce a NamedVector of given state, parameters, or optimized vars.

Setting entire state and parameter vectors

MTKHelpers.get_u_map — Function

get_u_map(names_u, pset::AbstractProblemParSetter)
get_p_map(names_p, pset::AbstractProblemParSetter)

Map each state and parameter the ProblemParSetter pset to a position in names.

When construction an ODEProblem from a ODESystem, the order of states and parameters may have changed compared with a previous construction.

In order to set entire state or parameter vectors, a mapping from current to previous positions, i.e. integer indices, is required, so that one can get a vectors in the new format by

u0_old[u_map]

p_old[p_map]

The mapping is constructed by supplying the names of u0old and pold to a ProblemParameterSetter constructed with the current ODESystem.

Keyword arguments

do_warn_missing: set to true to issue warnings if some ODESystem state or parameter names are not found in the old names. This may give false warnings for System parameters that have defaults and do not need to be part of the parameter vector.