Setting Initial state and Parameters of a problem

Oftern one wants to change a subset of the initial states,u0, and a subset of parameters,p, of an ODEProblem during an optimization.

Given u0 and p being a sequence of scalars, and popt is a sequence of optimized parameters, which may include initial states, the following class helps updating the corresponding positions in an ODEProblem.

Currently, parameters to optimize must be of the form u0...,p..., i.e. the position of all states to optimize must be first, before the parameters to optimize. Initial states and parameters must have different names.

# setting up a simple example composite system and problem
using ModelingToolkit, DifferentialEquations, LabelledArrays
using MTKHelpers
function samplesystem(;name,τ = 3.0, p1=1.1, p2=1.2)
    @variables t
    D = Differential(t)
    sts = @variables x(t) RHS(t)             # RHS is observed
    ps = @parameters τ=τ p1=p1 p2=p2       # parameters
    ODESystem([ RHS  ~ p1/p2 * (1 - x)/τ, D(x) ~ RHS ], t, sts, ps; name)
end
@named m = samplesystem()
@named sys = embed_system(m)
prob = ODEProblem(sys, [m.x => 0.0], (0.0,10.0))

# setup position matching
popt = SLVector(m₊x=0.1, m₊p1=2.1)
ps = ProblemParSetter(sys, keys(popt))

# extract optimized
get_paropt_labeled(ps, prob)

# update states and parameters
prob2 = update_statepar(ps, popt, prob)
get_paropt_labeled(ps, prob2) == popt

true

ProblemParSetter

The following type stores the necessary information, that can be queried.

MTKHelpers.ProblemParSetter — Type

ProblemParSetter(state_names,par_names,popt_names) 
ProblemParSetter(sys::ODESystem, popt_names; strip=false)

Helps keeping track of a subset of initial states and paraemters to be optimized.

Arguments

state_names: Tuple or AbstractVector of all the initial states of the problem

par_names: all the parameters of the problem

popt_names: the parameter/initial states to be optimized.

The states and parameters can be extracted from an ModelingToolkit.ODESystem. If strip=true, then namespaces of parameteres of a composed system are removed, e.g. subcomp₊p becomes p.

MTKHelpers.symbols_state — Method

symbols_state(ps::ProblemParSetter)
symbols_par(ps::ProblemParSetter)
symbols_paropt(ps::ProblemParSetter)

Report the names, i.e. symbols of problem states, problem parameters and optimized parameters respectively.

MTKHelpers.count_state — Function

count_state(::ProblemParSetter) 
count_par(::ProblemParSetter) 
count_paropt(::ProblemParSetter)

Report the number of problem states, problem parameters and optimized parameters respectively.

Updating a problem

MTKHelpers.update_statepar — Function

prob_new = update_statepar(ps::ProblemParSetter, popt, prob::ODEProblem) 
u0new, pnew = update_statepar(ps::ProblemParSetter, popt, u0, p)

Return an updated problem or updates states and parameters where values corresponding to positions in popt are set.

Further a variant or merge is implemented to produce an updated version of an SLArray or LArray are produced given a NamedTuple of updates.

using LabelledArrays, MTKHelpers
popt = SLVector(L = 10.1, k_L = 1.1, k_R = 1/20.1)
popt2 = merge(popt, (k_L = 1.2,))
popt2.k_L == 1.2

true

Extracting optimized parameters

MTKHelpers.get_paropt — Function

get_paropt(ps::ProblemParSetter, prob::ODEProblem; kwargs...)
get_paropt(ps::ProblemParSetter, u0, p)

get_paropt_labeled(ps::ProblemParSetter, prob::ODEProblem; kwargs...)
get_paropt_labeled(ps::ProblemParSetter, u0, p)

Extract the initial states and parameters corresponding to the positions that are optimized. If both u0 and p are AbstractVectors, the result is a Vector, otherwise the result is a Tuple.

The lebeled versions additionally call `labelparopt(see [label_state`](@ref)) on the return value.

Labeling

MTKHelpers.label_state — Function

label_state(ps, u::AbstractVector) = LArray{statesyms(ps)}(u)
label_par(ps, par::AbstractVector) = LArray{symbols_par(ps)}(par)
label_paropt(ps, popt::AbstractVector) = LArray{symbols_paropt(ps)}(popt)

Produce a labeled version of a sequence of initial states, parameters, or optimized parameters respectively. The return type differs given the input

SVector -> SLVector

NTuple -> NamedTuple

AbstractVector -> LArray