Translating symbols and Nums

ModelingToolkit constructs and ODEProblem from an ODESystem by supplying Dictionaries that map Nums to values. However, it is more convenient to store initial states and parameters as ComponentVectors with symbolic keys, instead of Dictionaries with Num keys.

The following functions help to translate symbols to Nums of the system.

# setting up a simple example composite system and problem
using ModelingToolkit, DifferentialEquations, ComponentArrays
using MTKHelpers
function samplesystem(;name,τ = 3.0, p1=1.1, p2=1.2)
    @variables t
    D = Differential(t)
    sts = @variables x(t) RHS(t)             # RHS is observed
    ps = @parameters τ=τ p1=p1 p2=p2       # parameters
    ODESystem([ RHS  ~ p1/p2 * (1 - x)/τ, D(x) ~ RHS ], t, sts, ps; name)
end
@named m = samplesystem()
@named sys = embed_system(m)

# setup initial conditions and parameters as ComponentVectors
u0 = ComponentVector(m₊x=0.0)
p = ComponentVector(m₊p1=1.11, m₊τ = 3.1,)

# convert to Dict(Num -> value) in order to create the ODEProblem
p_numdict = system_num_dict(p, m)
prob = ODEProblem(sys, system_num_dict(u0,m), (0,2), p_numdict);

pset = ProblemParSetter(sys, ComponentVector())
p_prob = label_par(pset, prob.p)
p_prob.m₊p2 = 1.2 # from default
p_prob.m₊τ == 3.1 # from p

true

system_num_dict

MTKHelpers.system_num_dict — Function

system_num_dict(d, sys::AbstractSystem, string_sys=sys.name)
system_num_dict(d, symbol_dict::AbstractDict)
system_num_dict(d, systems::NTuple)

Create a Dictionary Num=>value from symbolic Dictionary or ComponentVector.

Omit pairs where no Num was found.

In the third variant, a tuple of AbstractSystems can be specified to replace Nums of several (sub-)systems in the dictionary.

MTKHelpers.get_system_symbol_dict — Function

get_system_symbol_dict(sys::AbstractSystem, string_sys::String=string(sys.name))
get_system_symbol_dict(systems...)

Construct a Dict{Symbol => Num} for all properties in sys. All Symbols are prefixed with <string_sys>₊

The second variant merges the dictionaries obtained from several systems.